Removed 'fetch_latest.sh' from the Splunk app
Due to Splunk packaging standards for Splunk Cloud, the 'fetch_latest.sh' script
has been removed from the exported splunk app.  This script is still available
in the Git repo, it simply isn't included in the app uploaded to Splunk Base.
Instructions were added to the README documenting the steps to download the
script along with a brief explanation of the situation.